+++ sources/gdal/large_image_source_gdal/__init__.py
@@ -16,10 +16,13 @@
 
 import math
 import numpy
+import os
 import palettable
+import pathlib
 import PIL.Image
 import pyproj
 import struct
+import tempfile
 import threading
 from operator import attrgetter
 from osgeo import gdal
@@ -28,8 +31,11 @@ from osgeo import gdalconst
 from osgeo import osr
 from pkg_resources import DistributionNotFound, get_distribution
 
+import large_image
 from large_image.cache_util import LruCacheMetaclass, methodcache, CacheProperties
-from large_image.constants import SourcePriority, TileInputUnits, TILE_FORMAT_NUMPY, TILE_FORMAT_PIL
+from large_image.constants import (
+    SourcePriority, TileInputUnits, TileOutputMimeTypes,
+    TILE_FORMAT_NUMPY, TILE_FORMAT_PIL, TILE_FORMAT_IMAGE)
 from large_image.exceptions import TileSourceException
 from large_image.tilesource import FileTileSource
 
@@ -876,6 +882,31 @@ class GDALFileTileSource(FileTileSource, metaclass=LruCacheMetaclass):
         return super()._getRegionBounds(
             metadata, left, top, right, bottom, width, height, units, **kwargs)
 
+    def pixelToProjection(self, x, y, level=None):
+        """
+        Convert from pixels back to projection coordinates.
+
+        :param x, y: base pixel coordinates.
+        :param level: the level of the pixel.  None for maximum level.
+        :returns: x, y in projection coordinates.
+        """
+        if level is None:
+            level = self.levels - 1
+        if not self.projection:
+            x *= 2 ** (self.levels - 1 - level)
+            y *= 2 ** (self.levels - 1 - level)
+            gt = self._getGeoTransform()
+            px = gt[0] + gt[1] * x + gt[2] * y
+            py = gt[3] + gt[4] * x + gt[5] * y
+            return px, py
+        xScale = 2 ** level * self.tileWidth
+        yScale = 2 ** level * self.tileHeight
+        x = x / xScale - 0.5
+        y = 0.5 - y / yScale
+        x = x * self.unitsAcrossLevel0 + self.projectionOrigin[0]
+        y = y * self.unitsAcrossLevel0 + self.projectionOrigin[1]
+        return x, y
+
     @methodcache()
     def getThumbnail(self, width=None, height=None, **kwargs):
         """
@@ -971,6 +1002,124 @@ class GDALFileTileSource(FileTileSource, metaclass=LruCacheMetaclass):
                         except RuntimeError:
                             pass
         return pixel
+
+    def _encodeTiledImageFromVips(self, vimg, iterInfo, image, **kwargs):
+        """
+        Save a vips image as a tiled tiff.
+
+        :param vimg: a vips image.
+        :param iterInfo: information about the region based on the tile
+            iterator.
+        :param image: a record with partial vips images and the current output
+            size.
+
+        Additional parameters are available.
+
+        :param compression: the internal compression format.  This can handle
+            a variety of options similar to the converter utility.
+        :returns: a pathlib.Path of the output file and the output mime type.
+        """
+        convertParams = large_image.tilesource.base._vipsParameters(
+            defaultCompression='lzw', **kwargs)
+        convertParams.pop('pyramid', None)
+        vimg = large_image.tilesource.base._vipsCast(
+            vimg, convertParams['compression'] in {'webp', 'jpeg'})
+        gdalParams = large_image.tilesource.base._gdalParameters(
+            defaultCompression='lzw', **kwargs)
+        for ch in range(image['channels']):
+            gdalParams += [
+                '-b' if ch not in (1, 3) or ch + 1 != image['channels'] else '-mask', str(ch + 1)]
+        tl = self.pixelToProjection(
+            iterInfo['region']['left'], iterInfo['region']['top'], iterInfo['level'])
+        br = self.pixelToProjection(
+            iterInfo['region']['right'], iterInfo['region']['bottom'], iterInfo['level'])
+        gdalParams += [
+            '-a_srs',
+            iterInfo['metadata']['bounds']['srs'],
+            '-a_ullr',
+            str(tl[0]),
+            str(tl[1]),
+            str(br[0]),
+            str(br[1]),
+        ]
+        fd, tempPath = tempfile.mkstemp('.tiff', 'tiledRegion_')
+        os.close(fd)
+        fd, outputPath = tempfile.mkstemp('.tiff', 'tiledGeoRegion_')
+        os.close(fd)
+        try:
+            vimg.write_to_file(tempPath, **convertParams)
+            ds = gdal.Open(tempPath, gdalconst.GA_ReadOnly)
+            gdal.Translate(outputPath, ds, options=gdalParams)
+            os.unlink(tempPath)
+        except Exception as exc:
+            try:
+                os.unlink(tempPath)
+            except Exception:
+                pass
+            try:
+                os.unlink(outputPath)
+            except Exception:
+                pass
+            raise exc
+        return pathlib.Path(outputPath), TileOutputMimeTypes['TILED']
+
+    def getRegion(self, format=(TILE_FORMAT_IMAGE, ), **kwargs):
+        """
+        Get a rectangular region from the current tile source.  Aspect ratio is
+        preserved.  If neither width nor height is given, the original size of
+        the highest resolution level is used.  If both are given, the returned
+        image will be no larger than either size.
+
+        :param format: the desired format or a tuple of allowed formats.
+            Formats are members of (TILE_FORMAT_PIL, TILE_FORMAT_NUMPY,
+            TILE_FORMAT_IMAGE).  If TILE_FORMAT_IMAGE, encoding may be
+            specified.
+        :param kwargs: optional arguments.  Some options are region, output,
+            encoding, jpegQuality, jpegSubsampling, tiffCompression, fill.  See
+            tileIterator.
+        :returns: regionData, formatOrRegionMime: the image data and either the
+            mime type, if the format is TILE_FORMAT_IMAGE, or the format.
+        """
+        if not isinstance(format, (tuple, set, list)):
+            format = (format, )
+        # The tile iterator handles determining the output region
+        iterInfo = self._tileIteratorInfo(**kwargs)
+        # Only use gdal.Warp of the original image if the region has not been
+        # styled.
+        useGDALWarp = (
+            iterInfo and
+            not self._jsonstyle and
+            TILE_FORMAT_IMAGE in format and
+            kwargs.get('encoding') == 'TILED')
+        if not useGDALWarp:
+            return super().getRegion(format, **kwargs)
+        srs = self.projection or self.getProj4String()
+        tl = self.pixelToProjection(
+            iterInfo['region']['left'], iterInfo['region']['top'], iterInfo['level'])
+        br = self.pixelToProjection(
+            iterInfo['region']['right'], iterInfo['region']['bottom'], iterInfo['level'])
+        outWidth = iterInfo['output']['width']
+        outHeight = iterInfo['output']['height']
+        gdalParams = large_image.tilesource.base._gdalParameters(
+            defaultCompression='lzw', **kwargs)
+        gdalParams += [
+            '-t_srs', srs,
+            '-te', str(tl[0]), str(br[1]), str(br[0]), str(tl[1]),
+            '-ts', str(int(math.floor(outWidth))), str(int(math.floor(outHeight))),
+        ]
+        fd, outputPath = tempfile.mkstemp('.tiff', 'tiledGeoRegion_')
+        os.close(fd)
+        try:
+            self.logger.info('Using gdal warp %r' % gdalParams)
+            ds = gdal.Open(self._path, gdalconst.GA_ReadOnly)
+            gdal.Warp(outputPath, ds, options=gdalParams)
+        except Exception as exc:
+            try:
+                os.unlink(outputPath)
+            except Exception:
+                pass
+            raise exc
+        return pathlib.Path(outputPath), TileOutputMimeTypes['TILED']
 
 
 def open(*args, **kwargs):